+++ src/renderer/App.tsx
@@ -12,6 +12,7 @@ import {
   DiskStatus,
   StorageFilter,
   ActivityStatus,
+  KillVideoStatus,
 } from 'main/types';
 import Box from '@mui/material/Box';
 import { getLocalePhrase, Language } from 'localisation/translations';
@@ -36,6 +37,7 @@ import { Phrase } from 'localisation/phrases';
 import _ from 'lodash';
 import { playAudio } from './sounds';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+import KillVideoProgress from './KillVideoProgress';
 
 const ipc = window.electron.ipcRenderer;
 const queryClient = new QueryClient();
@@ -59,6 +61,11 @@ const WarcraftRecorder = () => {
   const [savingStatus, setSavingStatus] = useState<SaveStatus>(
     SaveStatus.NotSaving,
   );
+
+  const [killVideoStatus, setKillVideoStatus] = useState<KillVideoStatus>({
+    inProgress: false,
+    perc: 0,
+  });
 
   const [updateAvailable, setUpdateAvailable] = useState(false);
 
@@ -455,6 +462,7 @@ const WarcraftRecorder = () => {
         }}
       >
         <Toaster />
+        <KillVideoProgress language={appState.language} />
         <QueryClientProvider client={queryClient}>
           <TooltipProvider>
             <RendererTitleBar />